requirements.txt
datasets==2.14.6
evaluate==0.4.0
nltk==3.8.1
numpy==1.23.1
scikit-learn==1.2.2
scipy==1.10.1
seqeval==1.2.2
spams-bin==2.6.7
torch==2.0.1
torch-optimizer==0.3.0
tqdm==4.66.1
transformers==4.34.1
